rm unused stuff
