+++ server/handler.go
@@ -214,24 +214,3 @@ func (s *server) DeleteUserHandler(w http.ResponseWriter, r *http.Request) {
 		"user_id": userID,
 	})
 }
-
-func RespondWithError(w http.ResponseWriter, code int, message string) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(code)
-	json.NewEncoder(w).Encode(m.ErrorResponse{Code: code, Message: message})
-}
-
-func RespondWithSuccess(w http.ResponseWriter, code int, message string, data interface{}) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(code)
-
-	response := struct {
-		Message string      `json:"message"`
-		Data    interface{} `json:"data"`
-	}{
-		Message: message,
-		Data:    data,
-	}
-
-	json.NewEncoder(w).Encode(response)
-}
