+++ server/server.go
@@ -2,7 +2,9 @@ package server
 
 import (
 	"context"
+	"encoding/json"
 	s "microblogging/service"
+	"net/http"
 
 	"github.com/go-playground/validator"
 )
@@ -19,4 +21,25 @@ func NewServer(ctx context.Context, svc s.BlogService) *server {
 		Svc: svc,
 		ctx: ctx,
 	}
+}
+
+func RespondWithError(w http.ResponseWriter, code int, message string) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	json.NewEncoder(w).Encode(m.ErrorResponse{Code: code, Message: message})
+}
+
+func RespondWithSuccess(w http.ResponseWriter, code int, message string, data interface{}) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+
+	response := struct {
+		Message string      `json:"message"`
+		Data    interface{} `json:"data"`
+	}{
+		Message: message,
+		Data:    data,
+	}
+
+	json.NewEncoder(w).Encode(response)
 }
