+++ server/validators.go
@@ -2,8 +2,6 @@ package server
 
 import (
 	"errors"
-	m "microblogging/model"
-	"net/http"
 
 	"github.com/google/uuid"
 )
@@ -18,14 +16,4 @@ func ValidateContent(content string) error {
 func IsValidUUID(u string) bool {
 	_, err := uuid.Parse(u)
 	return err == nil
-}
-
-func ValidateUserID(userID string) (int, string) {
-	if userID == "" {
-		return http.StatusBadRequest, m.ErrMissingUserID.Error()
-	}
-	if !IsValidUUID(userID) {
-		return http.StatusBadRequest, m.ErrInvalidUUID.Error()
-	}
-	return http.StatusOK, ""
 }
